Performed refactoring on React Bootstrap scaffolding pages
