+++ lib/scaffold-react-project-structure.js
@@ -2,6 +2,7 @@
 const fs = require("fs");
 const path = require("path");
 const { createMUIPages } = require("./pages/mui");
+const { createBootstrapPages } = require("./pages/bootstrap");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -138,606 +139,6 @@ export const router = createBrowserRouter([
   } else {
     // Default to MUI if no specific UI is chosen
     createMUIPages(pagesDir, extx, ext, useTS);
-  }
-}
-
-function createBootstrapPages(pagesDir, extx, ext, useTS) {
-  const home = path.join(pagesDir, `Home.${extx}`);
-  if (!fs.existsSync(home)) {
-    writeFileSafe(home, `export default function Home() {
-  const features = [
-    { icon: '🚀', title: 'Performance', description: 'Lightning-fast builds with modern tooling', color: 'primary' },
-    { icon: '👥', title: 'Team Ready', description: 'Built for collaboration and scalability', color: 'success' },
-    { icon: '⚡', title: 'Fast Development', description: 'Hot reload and instant feedback', color: 'info' },
-    { icon: '🛡️', title: 'Type Safe', description: 'Full TypeScript support with strict mode', color: 'warning' },
-    { icon: '🧹', title: 'Clean Code', description: 'ESLint + Prettier for code quality', color: 'secondary' },
-    { icon: '🎨', title: 'Beautiful UI', description: 'Bootstrap components with theming', color: 'danger' }
-  ];
-
-  return (
-    <div className="container py-5">
-      <div className="text-center mb-5">
-        <h1 className="display-4 fw-bold text-primary mb-3">
-          Welcome to PolyFront
-        </h1>
-        <p className="lead text-muted mb-4">
-          Your comprehensive multi-frontend scaffold with everything you need
-        </p>
-        <p className="text-muted mx-auto" style={{ maxWidth: '600px' }}>
-          Built with React, TypeScript, Vite, Bootstrap, and modern development practices.
-          Start building your next great application today.
-        </p>
-      </div>
-
-      <div className="row g-4 mb-5">
-        {features.map((feature, index) => (
-          <div className="col-12 col-sm-6 col-md-4" key={index}>
-            <div className="card h-100 shadow-sm border-0">
-              <div className="card-body text-center">
-                <div className="display-4 mb-3">{feature.icon}</div>
-                <h5 className="card-title">{feature.title}</h5>
-                <p className="card-text text-muted">{feature.description}</p>
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-
-      <div className="card shadow-sm border-0">
-        <div className="card-body text-center p-5">
-          <h2 className="mb-3">Ready to Get Started?</h2>
-          <p className="text-muted mb-4">
-            This scaffold includes routing, state management, API integration, testing, and much more.
-            Explore the different pages to see all the features in action.
-          </p>
-          <div className="mt-4">
-            <button className="btn btn-primary btn-lg me-3">View Documentation</button>
-            <button className="btn btn-outline-primary btn-lg">Learn More</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-  }
-
-  // Create other Bootstrap pages
-  createOtherBootstrapPages(pagesDir, extx, ext, useTS);
-}
-
-function createOtherBootstrapPages(pagesDir, extx, ext, useTS) {
-  // About page
-  const about = path.join(pagesDir, `About.${extx}`);
-  if (!fs.existsSync(about)) {
-    writeFileSafe(about, `export default function About() {
-  const techStack = [
-    { name: 'React 19', version: 'Latest', category: 'Framework' },
-    { name: 'TypeScript', version: '5.x', category: 'Language' },
-    { name: 'Vite', version: '7.x', category: 'Build Tool' },
-    { name: 'Bootstrap', version: '5.x', category: 'UI Library' },
-    { name: 'React Router', version: '7.x', category: 'Routing' },
-    { name: 'Zustand', version: '5.x', category: 'State Management' },
-    { name: 'Axios', version: '1.x', category: 'HTTP Client' },
-    { name: 'Vitest', version: '3.x', category: 'Testing' }
-  ];
-
-  const features = [
-    'Modern React with hooks and functional components',
-    'Full TypeScript support with strict mode',
-    'Beautiful Bootstrap components with theming',
-    'Client-side routing with React Router',
-    'State management with Zustand',
-    'API integration with Axios',
-    'Comprehensive testing setup with Vitest',
-    'ESLint and Prettier for code quality',
-    'Hot module replacement for fast development',
-    'Responsive design with mobile-first approach'
-  ];
-
-  return (
-    <div className="container py-5">
-      <div className="text-center mb-5">
-        <h1 className="display-4 fw-bold text-primary mb-3">
-          About PolyFront
-        </h1>
-        <p className="lead text-muted mb-4">
-          A comprehensive scaffold for modern web applications
-        </p>
-      </div>
-
-      <div className="row g-4 mb-5">
-        <div className="col-md-6">
-          <div className="card h-100 shadow-sm">
-            <div className="card-body">
-              <h4 className="card-title text-primary">Technology Stack</h4>
-              <p className="card-text">
-                Built with the latest and greatest technologies in the React ecosystem.
-              </p>
-              <div className="row g-2">
-                {techStack.map((tech, index) => (
-                  <div className="col-6" key={index}>
-                    <span className="badge bg-primary me-2 mb-2">
-                      {tech.name} {tech.version}
-                    </span>
-                    <small className="text-muted d-block">{tech.category}</small>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="col-md-6">
-          <div className="card h-100 shadow-sm">
-            <div className="card-body">
-              <h4 className="card-title text-primary">Key Features</h4>
-              <ul className="list-unstyled">
-                {features.map((feature, index) => (
-                  <li key={index} className="mb-2">
-                    <i className="bi bi-check-circle-fill text-success me-2"></i>
-                    {feature}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div className="card shadow-sm">
-        <div className="card-body text-center p-5">
-          <h2 className="mb-4">Why Choose PolyFront?</h2>
-          <div className="row g-4">
-            <div className="col-md-4">
-              <div className="display-4 mb-3">⭐</div>
-              <h5>Production Ready</h5>
-              <p className="text-muted">Built with best practices and industry standards</p>
-            </div>
-            <div className="col-md-4">
-              <div className="display-4 mb-3">⚡</div>
-              <h5>Fast Development</h5>
-              <p className="text-muted">Hot reload and instant feedback for rapid iteration</p>
-            </div>
-            <div className="col-md-4">
-              <div className="display-4 mb-3">👥</div>
-              <h5>Team Friendly</h5>
-              <p className="text-muted">Consistent code style and comprehensive tooling</p>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-  }
-
-  // Dashboard page
-  const dashboard = path.join(pagesDir, `Dashboard.${extx}`);
-  if (!fs.existsSync(dashboard)) {
-    writeFileSafe(dashboard, `import { useState } from 'react';
-
-export default function Dashboard() {
-  const [stats] = useState({
-    users: 1234,
-    revenue: 45678,
-    orders: 89,
-    views: 12345
-  });
-
-  const [recentActivity] = useState([
-    { id: 1, user: 'Nirmal Samaranayaka', action: 'Order placed', time: '2 min ago', status: 'success' },
-    { id: 2, user: 'Jane Smith', action: 'Payment received', time: '5 min ago', status: 'success' },
-    { id: 3, user: 'Bob Johnson', action: 'Login attempt', time: '10 min ago', status: 'warning' },
-    { id: 4, user: 'Alice Brown', action: 'Account created', time: '15 min ago', status: 'success' },
-  ]);
-
-  const [topProducts] = useState([
-    { name: 'Premium Widget', sales: 156, revenue: 2340, growth: 12 },
-    { name: 'Super Gadget', sales: 98, revenue: 1890, growth: -5 },
-    { name: 'Mega Tool', sales: 87, revenue: 1450, growth: 8 },
-    { name: 'Ultra Device', sales: 76, revenue: 1230, growth: 15 },
-  ]);
-
-  const getStatusIcon = (status${useTS ? ':string' : ''}) => {
-    switch (status) {
-      case 'success': return <i className="bi bi-check-circle-fill text-success"></i>;
-      case 'warning': return <i className="bi bi-exclamation-triangle-fill text-warning"></i>;
-      case 'error': return <i className="bi bi-x-circle-fill text-danger"></i>;
-      default: return <i className="bi bi-check-circle-fill text-success"></i>;
-    }
-  };
-
-  const getGrowthColor = (growth ${useTS ? ': number' : ''}) => growth >= 0 ? 'success' : 'danger';
-
-  return (
-    <div>
-      <h1 className="display-4 fw-bold text-primary mb-3">Dashboard</h1>
-      <p className="lead text-muted mb-4">
-        Welcome back! Here's what's happening with your application today.
-      </p>
-
-      {/* Stats Cards */}
-      <div className="row g-3 mb-4">
-        <div className="col-12 col-sm-6 col-md-3">
-          <div className="card">
-            <div className="card-body">
-              <div className="d-flex align-items-center justify-content-between">
-                <div>
-                  <h6 className="text-muted mb-2">Total Users</h6>
-                  <h3 className="mb-2">{stats.users.toLocaleString()}</h3>
-                  <small className="text-success">
-                    <i className="bi bi-arrow-up me-1"></i>
-                    +12% from last month
-                  </small>
-                </div>
-                <div className="bg-primary text-white rounded-circle d-flex align-items-center justify-content-center" style={{ width: '56px', height: '56px' }}>
-                  <i className="bi bi-people-fill"></i>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="col-12 col-sm-6 col-md-3">
-          <div className="card">
-            <div className="card-body">
-              <div className="d-flex align-items-center justify-content-between">
-                <div>
-                  <h6 className="text-muted mb-2">Revenue</h6>
-                  <h3 className="mb-2">\${stats.revenue.toLocaleString()}</h3>
-                  <small className="text-success">
-                    <i className="bi bi-arrow-up me-1"></i>
-                    +8% from last month
-                  </small>
-                </div>
-                <div className="bg-success text-white rounded-circle d-flex align-items-center justify-content-center" style={{ width: '56px', height: '56px' }}>
-                  <i className="bi bi-currency-dollar"></i>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="col-12 col-sm-6 col-md-3">
-          <div className="card">
-            <div className="card-body">
-              <div className="d-flex align-items-center justify-content-between">
-                <div>
-                  <h6 className="text-muted mb-2">Orders</h6>
-                  <h3 className="mb-2">{stats.orders}</h3>
-                  <small className="text-success">
-                    <i className="bi bi-arrow-up me-1"></i>
-                    +15% from last month
-                  </small>
-                </div>
-                <div className="bg-warning text-white rounded-circle d-flex align-items-center justify-content-center" style={{ width: '56px', height: '56px' }}>
-                  <i className="bi bi-cart"></i>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="col-12 col-sm-6 col-md-3">
-          <div className="card">
-            <div className="card-body">
-              <div className="d-flex align-items-center justify-content-between">
-                <div>
-                  <h6 className="text-muted mb-2">Page Views</h6>
-                  <h3 className="mb-2">{stats.views.toLocaleString()}</h3>
-                  <small className="text-success">
-                    <i className="bi bi-arrow-up me-1"></i>
-                    +22% from last month
-                  </small>
-                </div>
-                <div className="bg-info text-white rounded-circle d-flex align-items-center justify-content-center" style={{ width: '56px', height: '56px' }}>
-                  <i className="bi bi-eye"></i>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Charts and Tables */}
-      <div className="row g-3">
-        <div className="col-md-8">
-          <div className="card">
-            <div className="card-header d-flex justify-content-between align-items-center">
-              <h5 className="mb-0">Top Products</h5>
-              <button className="btn btn-primary btn-sm">View All</button>
-            </div>
-            <div className="card-body">
-              <div className="table-responsive">
-                <table className="table">
-                  <thead>
-                    <tr>
-                      <th>Product</th>
-                      <th className="text-end">Sales</th>
-                      <th className="text-end">Revenue</th>
-                      <th className="text-end">Growth</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {topProducts.map((product) => (
-                      <tr key={product.name} className="border-b border-gray-100">
-                        <td className="py-3 px-4">{product.name}</td>
-                        <td className="py-3 px-4 text-right">{product.sales}</td>
-                        <td className="py-3 px-4 text-right">\${product.revenue.toLocaleString()}</td>
-                        <td className="py-3 px-4 text-right">
-                          <span className={\`px-2 py-1 rounded-full text-xs font-medium \${getGrowthColor(product.growth)}\`}>
-                            {product.growth >= 0 ? '+' : ''}{product.growth}%
-                          </span>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="col-md-4">
-          <div className="card">
-            <div className="card-header">
-              <h5 className="mb-0">Recent Activity</h5>
-            </div>
-            <div className="card-body">
-              <div className="list-group list-group-flush">
-                {recentActivity.map((activity) => (
-                  <div key={activity.id} className="list-group-item d-flex align-items-start">
-                    <div className="me-3">
-                      {getStatusIcon(activity.status)}
-                    </div>
-                    <div className="flex-grow-1">
-                      <div className="fw-bold">{activity.action}</div>
-                      <small className="text-muted">
-                        {activity.user} — {activity.time}
-                      </small>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
-  }
-
-  // Profile page
-  const profile = path.join(pagesDir, `Profile.${extx}`);
-  if (!fs.existsSync(profile)) {
-    writeFileSafe(profile, `import { useState } from 'react';
-
-export default function Profile() {
-  const [isEditing, setIsEditing] = useState(false);
-  const [profile] = useState({
-    name: 'Nirmal Samaranayaka',
-    email: 'nirmal.fullstack@gmail.com',
-    phone: '+46 (72) xxx-xxxx',
-    location: 'Stockholm, Sweden',
-    company: 'Scania AB.',
-    position: 'Senior Fullstack Developer',
-    education: 'Computer Science, University of Colombo',
-    website: 'https://dev.to/nirmalsamaranayaka',
-    bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.'
-  });
-
-  const [settings, setSettings] = useState({
-    emailNotifications: true,
-    pushNotifications: false,
-    darkMode: false,
-    twoFactorAuth: true
-  });
-
-  const [skills] = useState([
-    'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends', 
-    'TypeScript', 'JavaScript', 'HTML/CSS', 'jQuery', 'Bootstrap', 'Tailwind CSS',
-    '.NET 6/7/8', 'ASP.NET Core', 'C#', 'Web API', 'Microservices', 'Azure', 'AWS'
-  ]);
-
-  const handleSave = () => {
-    setIsEditing(false);
-  };
-
-  const handleCancel = () => {
-    setIsEditing(false);
-  };
-
-  const handleSettingChange = (setting${useTS ? ':string' : ''}) => (event${useTS ? ': React.ChangeEvent<HTMLInputElement>' : ''}) => {
-    setSettings(prev => ({
-      ...prev,
-      [setting]: event.target.checked
-    }));
-  };
-
-  return (
-    <div>
-      <div className="d-flex justify-content-between align-items-center mb-4">
-        <h1 className="display-4 fw-bold text-primary">Profile</h1>
-        <div>
-          {isEditing ? (
-            <>
-              <button className="btn btn-primary me-2" onClick={handleSave}>
-                <i className="bi bi-check me-1"></i>Save
-              </button>
-              <button className="btn btn-outline-secondary" onClick={handleCancel}>
-                <i className="bi bi-x me-1"></i>Cancel
-              </button>
-            </>
-          ) : (
-            <button className="btn btn-primary" onClick={() => setIsEditing(true)}>
-              <i className="bi bi-pencil me-1"></i>Edit Profile
-            </button>
-          )}
-        </div>
-      </div>
-
-      <div className="row g-4">
-        {/* Profile Header */}
-        <div className="col-12">
-          <div className="card">
-            <div className="card-body">
-              <div className="d-flex align-items-center gap-4">
-                <div className="bg-primary text-white rounded-circle d-flex align-items-center justify-content-center" style={{ width: '120px', height: '120px', fontSize: '3rem' }}>
-                  {profile.name.split(' ').map(n => n[0]).join('')}
-                </div>
-                <div className="flex-grow-1">
-                  <h2 className="mb-2">{profile.name}</h2>
-                  <h5 className="text-muted mb-3">{profile.position} at {profile.company}</h5>
-                  <p className="text-muted mb-3">{profile.bio}</p>
-                  <div className="d-flex gap-1 flex-wrap">
-                    {skills.map((skill) => (
-                      <span key={skill} className="badge bg-secondary me-1 mb-1">{skill}</span>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Profile Details */}
-        <div className="col-md-8">
-          <div className="card">
-            <div className="card-header">
-              <h5 className="mb-0">Personal Information</h5>
-            </div>
-            <div className="card-body">
-              <div className="row g-3">
-                <div className="col-sm-6">
-                  <label className="form-label">Full Name</label>
-                  <input
-                    type="text"
-                    className="form-control"
-                    value={profile.name}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-sm-6">
-                  <label className="form-label">Email</label>
-                  <input
-                    type="email"
-                    className="form-control"
-                    value={profile.email}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-sm-6">
-                  <label className="form-label">Phone</label>
-                  <input
-                    type="tel"
-                    className="form-control"
-                    value={profile.phone}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-sm-6">
-                  <label className="form-label">Location</label>
-                  <input
-                    type="text"
-                    className="form-control"
-                    value={profile.location}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-sm-6">
-                  <label className="form-label">Company</label>
-                  <input
-                    type="text"
-                    className="form-control"
-                    value={profile.company}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-sm-6">
-                  <label className="form-label">Position</label>
-                  <input
-                    type="text"
-                    className="form-control"
-                    value={profile.position}
-                    disabled={!isEditing}
-                  />
-                </div>
-                <div className="col-12">
-                  <label className="form-label">Bio</label>
-                  <textarea
-                    className="form-control"
-                    rows={3}
-                    value={profile.bio}
-                    disabled={!isEditing}
-                  ></textarea>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Settings */}
-        <div className="col-md-4">
-          <div className="card">
-            <div className="card-header">
-              <h5 className="mb-0">Settings</h5>
-            </div>
-            <div className="card-body">
-              <div className="form-check form-switch mb-3">
-                <input
-                  className="form-check-input"
-                  type="checkbox"
-                  checked={settings.emailNotifications}
-                  onChange={handleSettingChange('emailNotifications')}
-                />
-                <label className="form-check-label">Email Notifications</label>
-              </div>
-              <div className="form-check form-switch mb-3">
-                <input
-                  className="form-check-input"
-                  type="checkbox"
-                  checked={settings.pushNotifications}
-                  onChange={handleSettingChange('pushNotifications')}
-                />
-                <label className="form-check-label">Push Notifications</label>
-              </div>
-              <div className="form-check form-switch mb-3">
-                <input
-                  className="form-check-input"
-                  type="checkbox"
-                  checked={settings.darkMode}
-                  onChange={handleSettingChange('darkMode')}
-                />
-                <label className="form-check-label">Dark Mode</label>
-              </div>
-              <div className="form-check form-switch mb-3">
-                <input
-                  className="form-check-input"
-                  type="checkbox"
-                  checked={settings.twoFactorAuth}
-                  onChange={handleSettingChange('twoFactorAuth')}
-                />
-                <label className="form-check-label">Two-Factor Auth</label>
-              </div>
-            </div>
-          </div>
-
-          <div className="card mt-3">
-            <div className="card-header">
-              <h5 className="mb-0">Quick Actions</h5>
-            </div>
-            <div className="card-body">
-              <button className="btn btn-outline-primary w-100 mb-2">Download Resume</button>
-              <button className="btn btn-outline-primary w-100 mb-2">Share Profile</button>
-              <button className="btn btn-outline-danger w-100">Delete Account</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}`);
   }
 }
 
@@ -3572,7 +2973,7 @@ const Layout = () => {
           </Typography>
 
           <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-            <Chip label="v1.0.23" size="small" variant="outlined" />
+            <Chip label="v1.0.24" size="small" variant="outlined" />
             {/* Language selector using i18n */}
             <Select
               value={currentLang}
@@ -3717,7 +3118,7 @@ const Layout = () => {
             </button>
             <h4 className="mb-0">PolyFront</h4>
             <div className="d-flex align-items-center gap-2">
-              <span className="badge bg-secondary">v1.0.23</span>
+              <span className="badge bg-secondary">v1.0.24</span>
               
               {/* Language Selector */}
               <select
@@ -3811,7 +3212,7 @@ export default Layout;`);
                 {/* Version + Language + User */}
                 <div className="flex items-center gap-3">
                   {/* Version */}
-                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.23</span>
+                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.24</span>
       
                   {/* Language selector */}
                   <select
@@ -3920,7 +3321,7 @@ style={{ borderRight: 0 }}
         >
           PolyFront
         </Typography.Title>
-        <Tag bordered={false}>v1.0.23</Tag>
+        <Tag bordered={false}>v1.0.24</Tag>
         <Select
           size="small"
           value={currentLang}
@@ -4114,7 +3515,7 @@ function Layout() {
           <Box flex="1" />
 
           <HStack gap={2}>
-            <Badge>v1.0.23</Badge>
+            <Badge>v1.0.24</Badge>
 
             {/* Language Select (v3 controlled) */}
             <Select.Root
