+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@3nvs/polyfront-scaffold",
-  "version": "1.0.23",
+  "version": "1.0.24",
   "description": "Scaffold React (Vite or Webpack) and Angular apps with language/UI/store/testing options. Works on Windows/macOS/Linux.",
   "type": "commonjs",
   "license": "MIT",
